--- src/main/java/ImagePanel.java
@@ -1,8 +1,0 @@
-package src.main.java;
-
-import javax.swing.JPanel;
-
-public class ImagePanel extends JPanel
-{
-	ImagePanel() {}
-}
