+++ src/main/java/src/main/java/ImagePanel.java
@@ -1,0 +1,8 @@
+package src.main.java;
+
+import javax.swing.JPanel;
+
+public class ImagePanel extends JPanel
+{
+	ImagePanel() {}
+}
