+++ src/main/java/src/main/java/Settings.java
@@ -1,0 +1,19 @@
+package src.main.java;
+
+public class Settings //this is getting replaced with a File with these values stored in it
+{
+	public static int getDefaultWidth()
+	{
+		return 800;
+	}
+
+	public static int getDefaultHeight()
+	{
+		return 800;
+	}
+
+	public static String getVersion()
+	{
+		return "0.0.1";
+	}
+}
